+++ pages/enrollment/summary.js
@@ -30,7 +30,6 @@ export default function summary() {
   }
 
   return (
-    <section class="section notification">
       <div>
         <div>
           <Formik
@@ -70,6 +69,5 @@ export default function summary() {
           </Formik>
         </div>
       </div>
-    </section>
   );
 }