+++ pages/info/terms.js
@@ -11,7 +11,7 @@ export default function terms() {
 
 
        return (
-         <section class="section notification has-text-centered">
+         <div class="notification has-text-centered">
            <SectionTitle name="Terms and Conditions" />
            <div class="container box is-size-5-mobile">
              <div className={styles.terms}>
@@ -28,6 +28,6 @@ export default function terms() {
                </ol>
              </div>
            </div>
-         </section>
+         </div>
        );
 }